Optimize absolute() and normalize().
This short-circuits absolute() if the path is already absolute.
Detecting absolute paths is very fast, whereas join() is relatively
slow.

For normalize(), this runs a single-pass check over a path to see
whether it needs to be normalized at all. This check isn't free, but
it's a lot less costly than parsing and reserializing the path, and it
covers the common case.

See dart-lang/glob#24962

R=rnystrom@google.com

Review URL: https://codereview.chromium.org//1455003002 .